部分编译告警修改

+++ frameworks/native/pulseaudio/modules/abstract/module_loopback.c
@@ -944,8 +944,8 @@ static int SinkInputProcessMsgCb(pa_msgobject *obj, int code, void *data, int64_
             u->latency_snapshot.recv_counter = u->output_thread_info.recv_counter;
             u->latency_snapshot.loopback_memblockq_length = pa_memblockq_get_length(u->memblockq);
             /* Add content of render memblockq to sink latency */
-            u->latency_snapshot.sink_latency = pa_sink_get_latency_within_thread(u->sink_input->sink, true) +
-                pa_bytes_to_usec(length, &u->sink_input->sink->sample_spec);
+            u->latency_snapshot.sink_latency = (int64_t)(pa_sink_get_latency_within_thread(u->sink_input->sink, true) +
+                pa_bytes_to_usec(length, &u->sink_input->sink->sample_spec));
             u->latency_snapshot.sink_timestamp = pa_rtclock_now();
             return 0;
         }
